requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.3
attrs==22.2.0
Brotli==1.0.9
cchardet==2.1.7
cffi==1.15.1
charset-normalizer==3.1.0
e==1.4.5
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pycares==4.3.0
pycparser==2.21
python-dotenv==1.0.1
redis==5.0.6
unshorten-fast @ file:///home/rivado/unshorten-fast
yarl==1.8.2
